+++ src/features/addonList/AddOnList.test.tsx
@@ -32,50 +32,6 @@ import AddOnList from "./AddOnList";*/
 
 // TODO: Doesn't succeed because of loading screen when fetching from backend
 describe("AddOnList component", () => {
-  // Sample add-on data
-  const mockUsers: User[] = [
-    {
-      id: "user1",
-      name: "User One",
-      email: "userone@gmail.com"
-    },
-    {
-      id: "user2",
-      name: "User Two",
-      email: "usertwo@gmail.com"
-    },
-    {
-      id: "user3",
-      name: "User Three",
-      email: "userthree@gmail.com"
-    }
-  ];
-  const mockAuthors: Author[] = [
-    {
-      id: "author1",
-      user: mockUsers[1]
-    },
-    {
-      id: "author2",
-      user: mockUsers[2]
-    }
-  ];
-  // const mockAddOns: Addon[] = [
-  //   {
-  //     id: "1",
-  //     name: "Mock Addon 1",
-  //     summary: "Mock Summary 1",
-  //     category: AddonCategory.VISUALISATION,
-  //     author: mockAuthors[0]
-  //   },
-  //   {
-  //     id: "2",
-  //     name: "Mock Addon 2",
-  //     summary: "Mock Summary 2",
-  //     category: AddonCategory.VISUALISATION,
-  //     author: mockAuthors[1]
-  //   }
-  // ];
   it("temp", () => {
     expect(true).toEqual(true);
   });
