+++ src/pages/AddOnPage.test.tsx
@@ -11,15 +11,16 @@ import AddOnPage from "./AddOnPage";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
 import { Provider } from "react-redux";
 import { storeWithMockAddons } from "../utils/test-utils";
+import { addonList } from "../temp/tempAddons";
 
 describe("AddOnPage", () => {
   it("renders the add-on information when found", () => {
     const store = storeWithMockAddons();
 
-    // Render the Add-on Page starting at route "/addons/1" using the mocked state
+    // Render the Add-on Page starting at the route corresponding to an existing mock addon
     render(
       <Provider store={store}>
-        <MemoryRouter initialEntries={["/addons/1"]}>
+        <MemoryRouter initialEntries={[`/addons/${addonList[1].id}`]}>
           <Routes>
             <Route path="/addons/:id" element={<AddOnPage />} />
           </Routes>
@@ -28,8 +29,8 @@ describe("AddOnPage", () => {
     );
 
     // Assert wether the add-on page correctly retrieves the information from its parameters
-    expect(screen.getByText("Mock Addon 1")).toBeDefined();
-    expect(screen.getByText("Mock Summary 1")).toBeDefined();
+    expect(screen.getByText(addonList[1].name)).toBeDefined();
+    expect(screen.getByText(addonList[1].summary)).toBeDefined();
   });
   it('renders "Add-on not found" if the add-on is not in the store', () => {
     const store = storeWithMockAddons();
