+++ src/temp/tempAddons.ts
@@ -7,8 +7,8 @@
  */
 
 // Defines temporary add-ons. Will be removed when backend is connected
-import type { Addon } from "../../types/AddOnTypes";
-import { AddonCategory } from "../../types/AddOnTypes";
+import type { Addon } from "../types/AddOnTypes";
+import { AddonCategory } from "../types/AddOnTypes";
 import { authorList } from "./tempAuthors";
 
 const shortAddon: Addon = {
